Remove minimap and add bundle analyzer
