+++ package.json
@@ -4,6 +4,7 @@
   "license": "AGPL-3.0-or-later",
   "homepage": "https://sfucsss.org/",
   "private": true,
+  "sideEffects": true,
   "scripts": {
     "lint": "prettier --write .",
     "prebuild": "rm -Rf ./build/*",
@@ -22,9 +23,7 @@
     "@fortawesome/free-solid-svg-icons": "^6.6.0",
     "@fortawesome/react-fontawesome": "^0.2.2",
     "babel-loader": "^9.1.3",
-    "concurrently": "^8.2.2",
     "css-loader": "^6.10.0",
-    "html2canvas": "^1.4.1",
     "prettier": "^3.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -33,6 +32,7 @@
     "style-loader": "^3.3.4",
     "tailwindcss": "^3.4.8",
     "webpack": "^5.90.3",
+    "webpack-bundle-analyzer": "^4.10.2",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^5.0.4"
   },
