+++ public/index.html
@@ -10,6 +10,7 @@
       href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;700&display=swap"
       rel="stylesheet"
     />
+    <script defer="defer" src="/static/js/react.js"></script>
     <script defer="defer" src="/static/js/main.js"></script>
   </head>
   <body>
