+++ public/tech_fair/2024/tech_fair.js
@@ -7,43 +7,43 @@ function updateBanner() {
 
 // if a string is broken into an array, each element that is itself an array should be colored
 const FRONT_PAGE_MESSAGES = [
-  "// welcome to Tech Fair!",
+  '// welcome to Tech Fair!',
   "// located at SFU's Burnaby campus",
-  "// computing science is cool!",
-  "// computer science is cool!",
-  "// thousands of students!",
-  "// we might have cake",
-  "// on top of a mountain",
-  "// meet local vancouver companies!",
-  "// build your professional image!",
+  '// computing science is cool!',
+  '// computer science is cool!',
+  '// thousands of students!',
+  '// we might have cake',
+  '// on top of a mountain',
+  '// meet local vancouver companies!',
+  '// build your professional image!',
   //"// reasons you should join tech fair:\n// 1. to make connections\n// 2. it's fun!",
   //["gameobject.GetComponent<", ["Brain", "#4a7a96"], ">().LevelUp();"],
-  "long i = *(long*) &f;\ni = 0x5f3759df - (i >> 1);", // \n// hmmm, maybe not...",
-  "cout << \"n=\" << this->booths.size() << endl;",
-  "goto tech_fair;",
-  "let booth = new CompanyBooth(info);",
+  'long i = *(long*) &f;\ni = 0x5f3759df - (i >> 1);', // \n// hmmm, maybe not...",
+  'cout << "n=" << this->booths.size() << endl;',
+  'goto tech_fair;',
+  'let booth = new CompanyBooth(info);',
   //"for(booth in companyBooths) {\n\tSFU_CSSS.TechFair.create(booth);\n}",
-  "Tech Fair received signal SIGSEV, \nSegmentation fault.",
-  "// :)",
+  'Tech Fair received signal SIGSEV, \nSegmentation fault.',
+  '// :)',
   //"// the coolest event on campus",
-  "// static webpages rule!",
-  "// be there or be rectangle",
-  "adsflghja;dfhj;kldgsdaf\n:q!", //"adsflghja;dfhj;kldgsdaf\n~ \n:q!",
-  "(write (car tech_fair_events))",
-  "// run by the 【 SFU CSSS 】",
-  "// ↑↑↓↓←→←→BA\n",
+  '// static webpages rule!',
+  '// be there or be rectangle',
+  'adsflghja;dfhj;kldgsdaf\n:q!', //"adsflghja;dfhj;kldgsdaf\n~ \n:q!",
+  '(write (car tech_fair_events))',
+  '// run by the 【 SFU CSSS 】',
+  '// ↑↑↓↓←→←→BA\n',
   //"// ∀ companies ∃ a perfect one",
   //"[1]+  Stopped\tvi tech_fair.cpp\ncsss@CR:~/tf/$ git add *\ncsss@CR:~/tf/$ git commit -m \"stuff\"",
-  "// now with more colours!",
-  "// bigger than a bread box!",
-  "// Tech Fair: Where Ideas Ignite!\n// - ChatGPT",
-  "// Computing Science Student Society \n// (CSSS)",
+  '// now with more colours!',
+  '// bigger than a bread box!',
+  '// Tech Fair: Where Ideas Ignite!\n// - ChatGPT',
+  '// Computing Science Student Society \n// (CSSS)',
   //"if let Booth(info) = tech_fair {\n\t//TODO: this\n}",
-  "// TODO: make tech fair awesome",
-  "// run anually",
-  "csss@CR:~/tf/$ fg\nbash: fg: current: no such job",
-  "// meet your next co-op!",
-  "// lots of tech companies!",
+  '// TODO: make tech fair awesome',
+  '// run anually',
+  'csss@CR:~/tf/$ fg\nbash: fg: current: no such job',
+  '// meet your next co-op!',
+  '// lots of tech companies!',
   '<h1>SFU_CSSS.TechFair(2024);</h1>',
   '// Micah is awesome (for real)',
   '// Patrick is awesome!!',
