+++ src/_shared/react/Grid/Container/Container.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { classList } from '../../helpers';
+import { classList } from '../../../js';
 import './Container.css';
 
 /**
