+++ src/_shared/react/Icon/Email.js
@@ -1,21 +1,21 @@
 import React from 'react';
 
-import { classList } from '../helpers';
+import { classList } from '../../js';
 
 import './Icon.css';
 
 /**
- * Bullet icon.
+ * Email icon.
  * @param {object} props
  * @param {string} props.className - Additional class(es) for SVG icon
  * @param {string} props.style - Additional styles
  */
-export function Bullet(props) {
+export function Email(props) {
   const { className, style } = props;
 
   return (
     <svg
-      className={classList(['csss-icon', 'fill', className || ''])}
+      className={classList(['csss-icon', className || ''])}
       version="1.1"
       xmlns="http://www.w3.org/2000/svg"
       xmlnsXlink="http://www.w3.org/1999/xlink"
@@ -25,7 +25,10 @@ export function Bullet(props) {
       xmlSpace="preserve"
       style={style}
     >
-      <circle cx="16" cy="16" r="4" />
+      <rect x="4" y="8" width="24" height="16" />
+      <polyline points="4,8 16,20 28,8 " />
+      <line x1="12" y1="16" x2="4" y2="24" />
+      <line x1="20" y1="16" x2="28" y2="24" />
     </svg>
   );
 }
