+++ src/_shared/react/Icon/Expand.js
@@ -1,16 +1,16 @@
 import React from 'react';
 
-import { classList } from '../helpers';
+import { classList } from '../../js';
 
 import './Icon.css';
 
 /**
- * Profile icon.
+ * Expand (resize) icon.
  * @param {object} props
  * @param {string} props.className - Additional class(es) for SVG icon
  * @param {string} props.style - Additional styles
  */
-export function Profile(props) {
+export function Expand(props) {
   const { className, style } = props;
 
   return (
@@ -25,12 +25,10 @@ export function Profile(props) {
       xmlSpace="preserve"
       style={style}
     >
-      <path
-        d="M28,16c0,3.7-1.6,7-4.2,9.1c-0.9-5.2-4-9.1-7.8-9.1s-6.9,3.9-7.8,9.1C5.6,23,4,19.7,4,16C4,9.4,9.4,4,16,4
-        S28,9.4,28,16z"
-      />
-      <circle cx="16" cy="12" r="4" />
-      <path d="M23.8,25.1C21.7,26.9,19,28,16,28s-5.7-1.1-7.8-2.9c0.9-5.2,4-9.1,7.8-9.1S22.9,19.9,23.8,25.1z" />
+      <polyline points="12,4 4,4 4,12 " />
+      <polyline points="20,4 28,4 28,12 " />
+      <polyline points="4,20 4,28 12,28 " />
+      <polyline points="20,28 28,28 28,20 " />
     </svg>
   );
 }
