+++ src/_shared/react/Icon/File.js
@@ -1,21 +1,21 @@
 import React from 'react';
 
-import { classList } from '../helpers';
+import { classList } from '../../js';
 
 import './Icon.css';
 
 /**
- * White bullet icon.
+ * File icon.
  * @param {object} props
  * @param {string} props.className - Additional class(es) for SVG icon
  * @param {string} props.style - Additional styles
  */
-export function WhiteBullet(props) {
+export function File(props) {
   const { className, style } = props;
 
   return (
     <svg
-      className={classList(['csss-icon', className || ''])}
+      className={classList(['csss-icon fill', className || ''])}
       version="1.1"
       xmlns="http://www.w3.org/2000/svg"
       xmlnsXlink="http://www.w3.org/1999/xlink"
@@ -25,7 +25,7 @@ export function WhiteBullet(props) {
       xmlSpace="preserve"
       style={style}
     >
-      <circle cx="16" cy="16" r="8" />
+      <path d="M6,6v20c0,1.1,0.9,2,2,2h16c1.1,0,2-0.9,2-2V12l-8-8H8C6.9,4,6,4.9,6,6z" />
     </svg>
   );
 }
