+++ src/_shared/react/Icon/Link.js
@@ -1,21 +1,21 @@
 import React from 'react';
 
-import { classList } from '../helpers';
+import { classList } from '../../js';
 
 import './Icon.css';
 
 /**
- * Pause icon.
+ * Link (arrow) icon.
  * @param {object} props
  * @param {string} props.className - Additional class(es) for SVG icon
  * @param {string} props.style - Additional styles
  */
-export function Pause(props) {
+export function Link(props) {
   const { className, style } = props;
 
   return (
     <svg
-      className={classList(['csss-icon', 'fill', className || ''])}
+      className={classList(['csss-icon', className || ''])}
       version="1.1"
       xmlns="http://www.w3.org/2000/svg"
       xmlnsXlink="http://www.w3.org/1999/xlink"
@@ -25,8 +25,8 @@ export function Pause(props) {
       xmlSpace="preserve"
       style={style}
     >
-      <rect x="8" y="8" width="4" height="16" />
-      <rect x="20" y="8" width="4" height="16" />
+      <line x1="28" y1="4" x2="4" y2="28" />
+      <polyline points="12,4 28,4 28,20 " />
     </svg>
   );
 }
