+++ src/_shared/react/Icon/Moon.js
@@ -1,16 +1,16 @@
 import React from 'react';
 
-import { classList } from '../helpers';
+import { classList } from '../../js';
 
 import './Icon.css';
 
 /**
- * Hamburger menu icon.
+ * Moon icon.
  * @param {object} props
  * @param {string} props.className - Additional class(es) for SVG icon
  * @param {string} props.style - Additional styles
  */
-export function Hamburger(props) {
+export function Moon(props) {
   const { className, style } = props;
 
   return (
@@ -25,9 +25,8 @@ export function Hamburger(props) {
       xmlSpace="preserve"
       style={style}
     >
-      <line x1="4" y1="8" x2="28" y2="8" />
-      <line x1="4" y1="16" x2="28" y2="16" />
-      <line x1="4" y1="24" x2="28" y2="24" />
+      {/* eslint-disable-next-line max-len */}
+      <path d="M27.5,19.5C26,24.4,21.4,28,16,28C9.4,28,4,22.6,4,16c0-5.4,3.6-10,8.5-11.5C12.2,5.6,12,6.8,12,8c0,6.6,5.4,12,12,12C25.2,20,26.4,19.8,27.5,19.5z" />
     </svg>
   );
 }
