+++ src/_shared/react/Icon/Play.js
@@ -1,21 +1,21 @@
 import React from 'react';
 
-import { classList } from '../helpers';
+import { classList } from '../../js';
 
 import './Icon.css';
 
 /**
- * Arrow icon.
+ * Play icon.
  * @param {object} props
  * @param {string} props.className - Additional class(es) for SVG icon
  * @param {string} props.style - Additional styles
  */
-export function Arrow(props) {
+export function Play(props) {
   const { className, style } = props;
 
   return (
     <svg
-      className={classList(['csss-icon', className || ''])}
+      className={classList(['csss-icon', 'fill', className || ''])}
       version="1.1"
       xmlns="http://www.w3.org/2000/svg"
       xmlnsXlink="http://www.w3.org/1999/xlink"
@@ -25,7 +25,8 @@ export function Arrow(props) {
       xmlSpace="preserve"
       style={style}
     >
-      <polyline points="4,12 16,24 28,12 " />
+      {/* eslint-disable-next-line max-len */}
+      <path d="M26.5,15.1c0.7,0.4,0.7,1.3,0,1.7L19,21.2l-7.5,4.3c-0.7,0.4-1.5-0.1-1.5-0.9V16V7.3c0-0.8,0.8-1.2,1.5-0.9l7.5,4.3 L26.5,15.1z" />
     </svg>
   );
 }
