+++ src/_shared/react/Icon/index.js
