+++ src/_shared/react/VSCode/NavBar.css
