+++ src/_shared/react/VSCode/NavBar.js
