+++ src/_shared/react/VSCode/NavFolder.js
