+++ src/_shared/react/VSCode/NavItem.js
