+++ src/_shared/react/VSCode/index.js
