+++ src/_shared/react/index.js
@@ -3,4 +3,3 @@ export * as Flex from './Flex';
 export * as Grid from './Grid';
 export * as Icon from './Icon';
 export * as VSCode from './VSCode';
-export * as helpers from './helpers';
